Loadlinux.c : the function that setups atags move loadlinux.c
