+++ monitor/context.c
@@ -6,6 +6,9 @@
 #include "vmm.h"
 #include <print.h>
 #include <string.h>
+#if defined(LINUX_GUEST)
+#include "loadlinux.h"
+#endif
 
 #define NUM_GUEST_CONTEXTS		NUM_GUESTS_STATIC
 
@@ -33,219 +36,54 @@ static uint8_t _switch_locked = 0;  /* further switch request will be ignored if
 #define FIXUP_BMGUEST_LOAD_AT_GUEST1
 #endif
 
-#if defined(LINUX_GUEST)
-// uboot test start
-/* list of possible tags */
-#define ATAG_NONE       0x00000000
-#define ATAG_CORE       0x54410001
-#define ATAG_MEM        0x54410002
-#define ATAG_VIDEOTEXT  0x54410003
-#define ATAG_RAMDISK    0x54410004
-#define ATAG_INITRD2    0x54420005
-#define ATAG_SERIAL     0x54410006
-#define ATAG_REVISION   0x54410007
-#define ATAG_VIDEOLFB   0x54410008
-#define ATAG_CMDLINE    0x54410009
-/* structures for each atag */
-
-struct atag_header {
-        uint32_t size; /* length of tag in words including this header */
-        uint32_t tag;  /* tag type */
-};
-struct atag_core {
-        uint32_t flags;
-        uint32_t pagesize;
-        uint32_t rootdev;
-};
-
-struct atag_mem {
-        uint32_t     size;
-        uint32_t     start;
-};
-struct atag_serialnr {
-        uint32_t low;
-        uint32_t high;
-};
-
-struct atag_revision {
-        uint32_t rev;
-};
-struct atag_cmdline {
-        char    cmdline[1];
-};
-struct atag_ramdisk {
-        uint32_t flags;      /* bit 0 = load, bit 1 = prompt */
-        uint32_t size;       /* decompressed ramdisk size in _kilo_ bytes */
-        uint32_t start;      /* starting block of floppy-based RAM disk image */
-};
-struct atag_videotext {
-        uint8_t              x;           /* width of display */
-        uint8_t              y;           /* height of display */
-        uint16_t             video_page;
-        uint8_t              video_mode;
-        uint8_t              video_cols;
-        uint16_t             video_ega_bx;
-        uint8_t              video_lines;
-        uint8_t              video_isvga;
-        uint16_t             video_points;
-};
-struct atag_initrd2 {
-        uint32_t start;      /* physical start address */
-        uint32_t size;       /* size of compressed ramdisk image in bytes */
-};
-struct atag_videolfb {
-        uint16_t             lfb_width;
-        uint16_t             lfb_height;
-        uint16_t             lfb_depth;
-        uint16_t             lfb_linelength;
-        uint32_t             lfb_base;
-        uint32_t             lfb_size;
-        uint8_t              red_size;
-        uint8_t              red_pos;
-        uint8_t              green_size;
-        uint8_t              green_pos;
-        uint8_t              blue_size;
-        uint8_t              blue_pos;
-        uint8_t              rsvd_size;
-        uint8_t              rsvd_pos;
-};
-struct atag {
-        struct atag_header hdr;
-        union {
-                struct atag_core         core;
-                struct atag_mem          mem;
-                struct atag_videotext    videotext;
-                struct atag_ramdisk      ramdisk;
-                struct atag_initrd2      initrd2;
-                struct atag_serialnr     serialnr;
-                struct atag_revision     revision;
-                struct atag_videolfb     videolfb;
-                struct atag_cmdline      cmdline;
-        } u;
-};
-
-#define tag_next(t)     ((struct atag *)((uint32_t *)(t) + (t)->hdr.size))
-#define tag_size(type)  ((sizeof(struct atag_header) + sizeof(struct type)) >> 2)
-static struct atag *params; /* used to point at the current tag */
-
-static void
-setup_core_tag(void * address,long pagesize)
-{
-    params = (struct atag *)address;         /* Initialise parameters to start at given address */
-
-    params->hdr.tag = ATAG_CORE;            /* start with the core tag */
-    params->hdr.size = tag_size(atag_core); /* size the tag */
-    params->u.core.flags = 1;               /* ensure read-only */
-    params->u.core.pagesize = pagesize;     /* systems pagesize (4k) */
-    params->u.core.rootdev = 0;             /* zero root device (typicaly overidden from commandline )*/
-
-    params = tag_next(params);              /* move pointer to next tag */
-}
-
-static void setup_revision_tag(void)
-{
-    params->hdr.tag = ATAG_REVISION;
-    params->hdr.size = tag_size (atag_revision);
-    params->u.revision.rev = 0xcfdfdfdf;
-    params = tag_next (params);
-}
-
-char *strcpy(char *dest, const char *src)
-{
-    uint32_t i;
-    for (i = 0; src[i] != '\0'; ++i)
-        dest[i] = src[i];
-    dest[i] = '\0';
-    return dest;
-}
-
-static void
-setup_cmdline_tag(const char * line)
-{
-    int linelen = strlen(line);
-
-    if(!linelen)
-        return;                             /* do not insert a tag for an empty commandline */
-
-    params->hdr.tag = ATAG_CMDLINE;         /* Commandline tag */
-    params->hdr.size = (sizeof(struct atag_header) + linelen + 1 + 4) >> 2;
-
-    strcpy(params->u.cmdline.cmdline,line); /* place commandline into tag */
-
-    params = tag_next(params);              /* move pointer to next tag */
-}
-
-static void
-setup_mem_tag(uint32_t start, uint32_t len)
-{
-    printh("setup_mem_tag start :  %x len : %x\n", start, len);
-    params->hdr.tag = ATAG_MEM;             /* Memory tag */
-    params->hdr.size = tag_size(atag_mem);  /* size tag */
-
-    params->u.mem.start = start;            /* Start of memory area (physical address) */
-    params->u.mem.size = len;               /* Length of area */
-
-    params = tag_next(params);              /* move pointer to next tag */
-}
-
-static void
-setup_end_tag(void)
-{
-    params->hdr.tag = ATAG_NONE;            /* Empty tag ends list */
-    params->hdr.size = 0;                   /* zero length */
-}
-
-// uboot test end
-#endif
-
 #if defined (LINUX_GUEST)
 static void _hyp_guest0_copy_zimage(void)
 {
-	extern uint32_t guest_bin_start;
-	extern uint32_t guest_bin_end;
+    extern uint32_t guest_bin_start;
+    extern uint32_t guest_bin_end;
 
-	uint32_t *src = &guest_bin_start;
+    uint32_t *src = &guest_bin_start;
     uint32_t *end = &guest_bin_end;
-	uint32_t *dst = &guest_bin_start + (0x20008000/4);
+    uint32_t *dst = &guest_bin_start + (0x20008000/4);
 
-	HVMM_TRACE_ENTER();
+    HVMM_TRACE_ENTER();
 
-	uart_print("Copying guest0 image to guest1\n\r");
-	uart_print(" src:");uart_print_hex32((uint32_t)src); 
-	uart_print(" dst:");uart_print_hex32((uint32_t)dst); 
-	uart_print(" size:");uart_print_hex32( (uint32_t)(end - src) * sizeof(uint32_t));uart_print("\n\r");
+    uart_print("Copying guest0 image to guest1\n\r");
+    uart_print(" src:");uart_print_hex32((uint32_t)src); 
+    uart_print(" dst:");uart_print_hex32((uint32_t)dst); 
+    uart_print(" size:");uart_print_hex32( (uint32_t)(end - src) * sizeof(uint32_t));uart_print("\n\r");
 
-	while(src < end ) {
-		*dst++ = *src++;
-	}
-	uart_print("=== done ===\n\r");
-	HVMM_TRACE_EXIT();
+    while(src < end ) {
+        *dst++ = *src++;
+    }
+    uart_print("=== done ===\n\r");
+    HVMM_TRACE_EXIT();
 }
 #endif
 
 #if defined (FIXUP_BMGUEST_LOAD_AT_GUEST1) 
 static void _hyp_fixup_unloaded_guest(void)
 {
-	extern uint32_t guest_bin_start;
-	extern uint32_t guest_bin_end;
+    extern uint32_t guest_bin_start;
+    extern uint32_t guest_bin_end;
     extern uint32_t guest2_bin_start;
 
-	uint32_t *src = &guest_bin_start;
+    uint32_t *src = &guest_bin_start;
     uint32_t *end = &guest_bin_end;
-	uint32_t *dst = &guest2_bin_start;
+    uint32_t *dst = &guest2_bin_start;
 
-	HVMM_TRACE_ENTER();
+    HVMM_TRACE_ENTER();
 
-	uart_print("Copying guest0 image to guest1\n\r");
-	uart_print(" src:");uart_print_hex32((uint32_t)src); 
-	uart_print(" dst:");uart_print_hex32((uint32_t)dst); 
+    uart_print("Copying guest0 image to guest1\n\r");
+    uart_print(" src:");uart_print_hex32((uint32_t)src); 
+    uart_print(" dst:");uart_print_hex32((uint32_t)dst); 
 	uart_print(" size:");uart_print_hex32( (uint32_t)(end - src) * sizeof(uint32_t));uart_print("\n\r");
 
-	while(src < end ) {
-		*dst++ = *src++;
-	}
-	uart_print("=== done ===\n\r");
-	HVMM_TRACE_EXIT();
+    while(src < end ) {
+        *dst++ = *src++;
+    }
+    uart_print("=== done ===\n\r");
+    HVMM_TRACE_EXIT();
 }
 #endif
 
@@ -480,7 +318,7 @@ void context_restore_cops(struct arch_regs_cop *regs_cop)
 
 
 /* DEPRECATED: use context_switchto(vmid) and context_perform_switch() 
-    void context_switch_to_next_guest(struct arch_regs *regs_current)
+   void context_switch_to_next_guest(struct arch_regs *regs_current)
  */
 
 static hvmm_status_t context_perform_switch_to_guest_regs(struct arch_regs *regs_current, vmid_t next_vmid)
@@ -601,18 +439,6 @@ void context_switch_to_initial_guest(void)
     context_perform_switch();
 }
 
-#if defined(LINUX_GUEST)
-static void setup_tags(uint32_t *src)
-{
-    char *commandline = "root=/dev/ram rw earlyprintk console=ttyAMA0 mem=256M rdinit=/sbin/init";
-    setup_core_tag(src+(0x100/4), 4096);       /* standard core tag 4k pagesize */
-    setup_cmdline_tag(commandline);    /* commandline setting root device */
-    setup_revision_tag();
-    setup_mem_tag((uint32_t)(src-(0x20000000/4)), 0x10000000);
-    /* end of tags */
-    setup_end_tag();
-}
-#endif
 
 void context_init_guests(void)
 {
