+++ monitor/string.c
@@ -135,3 +135,12 @@ char *(strchr)(char const *s, int c)
 	else
 		return (char *)0;
 }
+
+char *(strcpy)(char *dest, const char *src)
+{
+    unsigned int i;
+    for (i = 0; src[i] != '\0'; ++i)
+        dest[i] = src[i];
+    dest[i] = '\0';
+    return dest;
+}
