+++ monitor/string.h
@@ -12,5 +12,6 @@ extern int (memcmp)(void const *p1, void const *p2, size_t n);
 extern int (strcmp)(char const *s1, char const *s2);
 extern int (strncmp)(char const *s1, char const *s2, size_t n);
 extern char *(strchr)(char const *s, int c);
+extern char *(strcpy)(char *dest, const char *src);
 
 #endif
